% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_asr_enrollment}
\alias{download_asr_enrollment}
\title{Download enrollment data from Annual Statistical Reports}
\usage{
download_asr_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with district enrollment data
}
\description{
Downloads the Excel file from the Arkansas Division of Elementary and
Secondary Education's Annual Statistical Reports page.
}
\keyword{internal}
