% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_ade_fiscal_year_id}
\alias{get_ade_fiscal_year_id}
\title{Get the fiscal year ID for ADE Data Center}
\usage{
get_ade_fiscal_year_id(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Fiscal year ID string
}
\description{
The ADE Data Center uses fiscal year IDs in its postback system.
This function maps school year end to the fiscal year ID.
}
\keyword{internal}
