% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_district_id}
\alias{is_district_id}
\title{Check if ID is a district (not a school)}
\usage{
is_district_id(id)
}
\arguments{
\item{id}{Character ID}
}
\value{
Logical TRUE if this is a district ID
}
\description{
District IDs in Arkansas end in "000"
}
\keyword{internal}
