% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_school_year}
\alias{parse_school_year}
\title{Convert school year string to end year}
\usage{
parse_school_year(sy)
}
\arguments{
\item{sy}{School year string (e.g., "2023-2024" or "2023-24")}
}
\value{
Integer end year
}
\description{
Converts Arkansas school year format (e.g., "2023-2024") to end year (2024)
}
\keyword{internal}
